+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 # v2021.7.1-beta
 - bugfix - fix jslint not warning about function-redefinition when function is defined inside a call.
 - bugfix - fix website crashing when linting pure json-object.
+- jslint - add new beta-warning if functions are unordered.
 - jslint - add new warning disallowing string-literal as property-name, e.g. {`aa`:0}.
 - jslint - comment out shebang in jslint.mjs so older ios devices can use website.
 - tests - revamp cause-based testing with more robust instrumentation.